move progress and cancellation of file operation participants into its only customer so that progress stop when applying a workspace edit and showing its preview
